requirements.txt
torch 
torchvision 
torchaudio
numpy
argparse
natsorted
